Generate data panel into a specific library Fixes #340 - breakdow of gen_angular
