+++ test/ng/projects/test/src/lib/bstruct-db.ts
@@ -5,6 +5,9 @@ import { AstructDB } from './astruct-db'
 import { NullInt64 } from './null-int64'
 
 export class BstructDB {
+
+	static GONGSTRUCT_NAME = "Bstruct"
+
 	CreatedAt?: string
 	DeletedAt?: string
 	ID: number = 0
