--- test/ng/projects/testspecific/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/testspecific",
-  "lib": {
-    "entryFile": "src/public-api.ts"
-  }
-}
