--- test/ng/projects/testspecific/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "testspecific",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^15.2.0",
-    "@angular/core": "^15.2.0"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "sideEffects": false
-}
