--- test/ng/projects/testspecific/src/lib/testspecific.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'lib-testspecific',
-  template: `
-    <p>
-      testspecific works!
-    </p>
-  `,
-  styles: [
-  ]
-})
-export class TestspecificComponent {
-
-}
