--- test/ng/projects/testspecific/src/lib/testspecific.module.ts
@@ -1,38 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { TestspecificComponent } from './testspecific.component';
-import { DataModelPanelComponent } from './data-model-panel/data-model-panel.component';
-
-import { GongdocModule } from 'gongdoc'
-import { GongdocdiagramsModule } from 'gongdocdiagrams'
-
-import { TestModule } from 'test'
-
-import { MatRadioModule } from '@angular/material/radio';
-import { FormsModule } from '@angular/forms';
-import { CommonModule } from '@angular/common';
-import { AngularSplitModule } from 'angular-split';
-
-@NgModule({
-  declarations: [
-    TestspecificComponent,
-    DataModelPanelComponent
-  ],
-  imports: [
-    GongdocModule,
-    GongdocdiagramsModule,
-
-    MatRadioModule,
-    FormsModule,
-    CommonModule,
-
-    AngularSplitModule,
-
-    TestModule,
-  ],
-  exports: [
-    TestspecificComponent,
-    DataModelPanelComponent
-  ]
-})
-export class TestspecificModule { }
