--- test/ng/projects/testspecific/src/lib/testspecific.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class TestspecificService {
-
-  constructor() { }
-}
