--- test/ng/projects/testspecific/src/public-api.ts
@@ -1,7 +1,0 @@
-/*
- * Public API Surface of testspecific
- */
-
-export * from './lib/testspecific.service';
-export * from './lib/testspecific.component';
-export * from './lib/testspecific.module';
